+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changelog
 
+next
+    - cache columns and tables in the query builder
+
 2.2.2  2025-08-18
     - add support for uniq() aggregation function
     - fix setting limit from variables query